+++ .coverage
@@ -1,1 +1,1 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-2\\app\\__init__.py":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,16,19,20,21,23],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-2\\app\\api\\__init__.py":[1,3,4,5,8,9,10],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-2\\app\\api\\products.py":[1,2,3,4,5,7,9,11,16,17,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,69,71,72,73,75,78,79,80,81,82,83,84,85,86,88,90,91,92,95,96,99,100],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-2\\app\\api\\sales.py":[1,2,3,4,5,6,8,10,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,52,53],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-2\\app\\api\\registration.py":[1,2,3,4,5,7,9,12,14,15,16,17,18,19,20,21,22,24,25,26,27,28,29,30,31,32,35,36,37]}}
+!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-2\\app\\__init__.py":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,16,19,20,21,23],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-2\\app\\api\\__init__.py":[1,3,4,5,8,9,10],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-2\\app\\api\\products.py":[1,2,3,4,5,7,9,11,16,17,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,69,71,72,73,75,78,79,80,81,82,83,84,85,86,88,90,91,92,95,96,99,100],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-2\\app\\api\\sales.py":[1,2,3,4,5,6,8,10,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,52,53],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-2\\app\\api\\registration.py":[1,2,3,4,5,7,9,12,14,15,16,17,18,19,20,21,22,24,25,26,27,28,29,30,31,32,35,36,37],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-3\\app\\__init__.py":[1,2,3,4,5,6,7,10,11,12,13,14,15,17,23,24,25,27],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-3\\app\\registration\\__init__.py":[1,3,5],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-3\\app\\registration\\views.py":[1,2,3,4,5,6,8,10,15,17,40,41,63,64,42,43,44,46,49,52,53,55,56,59,60,18,19,20,21,22,25,28,30,32,35,36,37,29,31],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-3\\app\\products\\__init__.py":[1,3,5],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-3\\app\\products\\views.py":[1,2,3,4,5,6,8,10,13,14,27,41,56,62,29,30,31,32,33,35,37,38,39,36,16,22,23,25,17,18,20,19,34],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-3\\app\\sales\\__init__.py":[1,3,5],"C:\\Users\\Home\\Documents\\Development\\Andela Prep\\Store-Manager\\Store-Manager-Challenge-3\\app\\sales\\views.py":[1,2,3,4,5,6,7,9,12,15,16,18,26,41,28,29,30,19,23,24,20,21]}}
