+++ .travis.yml
@@ -7,6 +7,11 @@ python:
 install:
   - "pip install -r requirements.txt"
 
+services:
+  - postgresql
+
+before_script:
+  - psql -c 'CREATE DATABASE test_store_manager;' -U postgres
 
 # command to run tests
 script:
